# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Trading Social Media App",
        "version": "0.1.0"
    },
    "paths": {
        "/signal/service/api/v1/create-signal/": {
            "post": {
                "tags": [
                    "Signal Service"
                ],
                "summary": "Create Signal",
                "operationId": "create_signal_signal_service_api_v1_create_signal__post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_create_signal_signal_service_api_v1_create_signal__post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignalResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/signal/service/api/v1/get-all-signals/": {
            "get": {
                "tags": [
                    "Signal Service"
                ],
                "summary": "Get Signals",
                "operationId": "get_signals_signal_service_api_v1_get_all_signals__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/SignalResponse"
                                    },
                                    "type": "array",
                                    "title": "Response Get Signals Signal Service Api V1 Get All Signals  Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/signal/service/api/v1/get-user-signals/{user_id}/": {
            "get": {
                "tags": [
                    "Signal Service"
                ],
                "summary": "Get User Signals",
                "operationId": "get_user_signals_signal_service_api_v1_get_user_signals__user_id___get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SignalResponse"
                                    },
                                    "title": "Response Get User Signals Signal Service Api V1 Get User Signals  User Id   Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/signal/service/api/v1/get-signal/{signal_id}/": {
            "get": {
                "tags": [
                    "Signal Service"
                ],
                "summary": "Get Signal",
                "operationId": "get_signal_signal_service_api_v1_get_signal__signal_id___get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "signal_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Signal Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignalResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/signal/service/api/v1/update-signal/{signal_id}/": {
            "patch": {
                "tags": [
                    "Signal Service"
                ],
                "summary": "Update Signal",
                "operationId": "update_signal_signal_service_api_v1_update_signal__signal_id___patch",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "signal_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Signal Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSignalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/signal/service/api/v1/delete-signal/{signal_id}/": {
            "delete": {
                "tags": [
                    "Signal Service"
                ],
                "summary": "Delete Signal Obj",
                "operationId": "delete_signal_obj_signal_service_api_v1_delete_signal__signal_id___delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "signal_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Signal Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/signal/service/api/v1/get-signal-picture/{filename}/": {
            "get": {
                "tags": [
                    "Signal Service"
                ],
                "summary": "Get Picture",
                "operationId": "get_picture_signal_service_api_v1_get_signal_picture__filename___get",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Filename"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_create_signal_signal_service_api_v1_create_signal__post": {
                "properties": {
                    "short_or_long": {
                        "type": "string",
                        "title": "Short Or Long"
                    },
                    "signal_picture": {
                        "type": "string",
                        "format": "binary",
                        "title": "Signal Picture"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "signal_for_coin": {
                        "type": "string",
                        "title": "Signal For Coin"
                    },
                    "when": {
                        "type": "string",
                        "format": "date-time",
                        "title": "When"
                    },
                    "when_end": {
                        "type": "string",
                        "format": "date-time",
                        "title": "When End"
                    }
                },
                "type": "object",
                "required": [
                    "short_or_long",
                    "signal_picture",
                    "title",
                    "signal_for_coin",
                    "when",
                    "when_end"
                ],
                "title": "Body_create_signal_signal_service_api_v1_create_signal__post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SUser": {
                "properties": {
                    "id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Id"
                    },
                    "username": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Username"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "token": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Token"
                    },
                    "picture_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Picture Url"
                    }
                },
                "type": "object",
                "title": "SUser"
            },
...
            "SignalResponse": {
                "properties": {
                    "id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Id"
                    },
                    "title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Title"
                    },
                    "signal_picture": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Signal Picture"
                    },
                    "signal_for_coin": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Signal For Coin"
                    },
                    "created_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Created At"
                    },
                    "when": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "When"
                    },
                    "when_end": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "When End"
                    },
                    "short_or_long": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Short Or Long"
                    },
                    "user_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Id"
                    },
                    "user": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/SUser"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "title": "SignalResponse"
            },
            "UpdateSignalRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "signal_for_coin": {
                        "type": "string",
                        "title": "Signal For Coin"
                    },
                    "when": {
                        "type": "string",
                        "format": "date-time",
                        "title": "When"
                    },
                    "when_end": {
                        "type": "string",
                        "format": "date-time",
                        "title": "When End"
                    },
                    "short_or_long": {
                        "type": "string",
                        "title": "Short Or Long"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "signal_for_coin",
                    "when",
                    "when_end",
                    "short_or_long"
                ],
                "title": "UpdateSignalRequest"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "http://localhost:8000/auth-login/"
                    }
                }
            }
        }
    }
}